Added search engine, updated a few views

+++ RetroGamingWorld/Controllers/CommentsController.cs
@@ -1,6 +1,8 @@
 ﻿using RetroGamingWorld.Models;
 using Microsoft.AspNetCore.Mvc;
 using RetroGamingWorld.Data;
+using Microsoft.EntityFrameworkCore;
+using System.Runtime.CompilerServices;
 
 namespace RetroGamingWorld.Controllers
 {
@@ -29,36 +31,39 @@ namespace RetroGamingWorld.Controllers
         [HttpPost]
         public IActionResult Delete(int id)
         {
-            Comment comm = db.Comments.Find(id);
-            db.Comments.Remove(comm);
-            db.SaveChanges();
-            return Redirect("/Articles/Show/" + comm.ArticleId);
+            Comment? comm = db.Comments.Find(id);
+            if (comm is not null)
+            {
+                db.Comments.Remove(comm);
+                db.SaveChanges();
+                return Redirect("/Articles/Show/" + comm.ArticleId);
+            }
+            return NotFound();
         }
+        [HttpGet]
         public IActionResult Edit(int id)
         {
-            Comment comm = db.Comments.Find(id);
-            ViewBag.Comment = comm;
-            return View();
+            Comment? comm = db.Comments.Include(c => c.Article)
+                                .FirstOrDefault(c => c.Id == id);
+            Console.WriteLine(comm);
+            if (comm is null)
+                return NotFound();
+
+            return View(comm);
         }
 
         [HttpPost]
         public IActionResult Edit(int id, Comment requestComment)
         {
-            Comment comm = db.Comments.Find(id);
+            Comment? comm = db.Comments.Find(id);
             if (ModelState.IsValid)
             {
-
+                comm.Date = DateTime.Now;
                 comm.Content = requestComment.Content;
 
                 db.SaveChanges();
-
-                return Redirect("/Articles/Show/" + comm.ArticleId);
             }
-            else
-            {
-                return Redirect("/Articles/Show/" + comm.ArticleId);
-            }
-
+            return Redirect("/Articles/Show/" + comm.ArticleId);
         }
     }
 }
